MAJOR REORGANIZATION OF COMPONENTS AND CODEBASE
