+++ docs/README.md
@@ -32,4 +32,5 @@ Contiene toda la documentación relacionada con el proceso de desarrollo, incluy
 
 ## Estado Actual
 ✅ **Fase 1 completada**: Limpieza de archivos obsoletos
-⏳ **Próximo**: Fase 2 - Reorganización de componentes por funcionalidad
+✅ **Fase 2 completada**: Reorganización de componentes por funcionalidad
+⏳ **Próximo**: Fase 3 - Optimización de servicios y hooks
