+++ src/components/common/forms/index.ts
@@ -1,0 +1,2 @@
+// Common Form Components
+export { AvatarUpload } from './AvatarUpload'
