+++ src/components/common/media/ImageCarousel.tsx
