+++ src/components/common/media/ZoomableImage.tsx
