+++ src/components/common/media/index.ts
@@ -1,0 +1,3 @@
+// Common Media Components
+export { ZoomableImage } from './ZoomableImage'
+export { ImageCarousel } from './ImageCarousel'
