+++ src/components/common/ui/FileInfo.tsx
