+++ src/components/features/categories/CategoryManager.tsx
