+++ src/components/features/categories/CategoryManagerCompact.tsx
