+++ src/components/features/categories/CategorySelector.tsx
