+++ src/components/features/categories/index.ts
@@ -1,0 +1,5 @@
+// Category Management Components
+export { default as CategoryFilter } from './CategoryFilter'
+export { default as CategoryManager } from './CategoryManager'
+export { default as CategoryManagerCompact } from './CategoryManagerCompact'
+export { default as CategorySelector } from './CategorySelector'
