+++ src/components/features/links/SortableLinksForm/PreviewModalUpdated.tsx
@@ -1,9 +1,9 @@
 import Image from 'next/image'
 import { useState } from 'react'
-import { SocialIcon } from '@/app/components/SocialIcon'
-import { ProductModal } from '@/components/ProductModal'
-import { ItemModal } from '@/components/ItemModal'
-import { PressablesList } from '@/components/PressablesList'
+import { SocialIcon } from '../../../common/ui/SocialIcon'
+import { ProductModal } from '../../products/ProductModal'
+import { ItemModal } from '../../products/ItemModal'
+import { PressablesList } from '../../products/PressablesList'
 import { type ProfileImageType } from '@/utils/file-utils'
 import { LinkItem, Product, Item } from '@/types/product'
 import { StyleSettings } from '@/utils/links'
