+++ src/components/features/products/ItemForm.tsx
@@ -1,29 +1,32 @@
 'use client'
 
 import { useState, useEffect } from 'react'
-import { Product } from '@/types/product'
-import { ImageCarousel } from './ImageCarousel'
+import { Item } from '@/types/product'
+import { ImageCarousel } from '../../common/media/ImageCarousel'
 import { TrashIcon, PlusIcon, PhotoIcon } from '@heroicons/react/24/outline'
-import CategorySelector from '@/components/CategorySelector'
+import CategorySelector from '../categories/CategorySelector'
+import { EyeIcon, EyeSlashIcon } from '@heroicons/react/24/outline'
 
-interface ProductFormProps {
-    product?: Product
+interface ItemFormProps {
+    item?: Item
     availableCategories: string[]
-    onUpdate: (product: Partial<Product>) => void
+    onUpdate: (item: Partial<Item>) => void
     onRemove: () => void
 }
 
-export function ProductForm({ product, availableCategories, onUpdate, onRemove }: ProductFormProps) {
+export function ItemForm({ item, availableCategories, onUpdate, onRemove }: ItemFormProps) {
     const [errors, setErrors] = useState<Record<string, string>>({})
 
     // Estados locales para los campos
-    const [title, setTitle] = useState(product?.title || '')
-    const [subtitle, setSubtitle] = useState(product?.subtitle || '')
-    const [price, setPrice] = useState(product?.price?.toString() || '')
-    const [paymentLink, setPaymentLink] = useState(product?.paymentLink || '')
-    const [description, setDescription] = useState(product?.description || '')
-    const [images, setImages] = useState<string[]>(product?.images || [])
-    const [categories, setCategories] = useState<string[]>(product?.categories || [])
+    const [title, setTitle] = useState(item?.title || '')
+    const [subtitle, setSubtitle] = useState(item?.subtitle || '')
+    const [price, setPrice] = useState(item?.price?.toString() || '')
+    const [priceVisible, setPriceVisible] = useState(item?.priceVisible ?? true)
+    const [buttonText, setButtonText] = useState(item?.buttonText || 'Ver más')
+    const [paymentLink, setPaymentLink] = useState(item?.paymentLink || '')
+    const [description, setDescription] = useState(item?.description || '')
+    const [images, setImages] = useState<string[]>(item?.images || [])
+    const [categories, setCategories] = useState<string[]>(item?.categories || [])
 
     // Validaciones
     const validateField = (field: string, value: any) => {
@@ -60,9 +63,19 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                 }
                 break
 
+            case 'buttonText':
+                if (!value.trim()) {
+                    newErrors.buttonText = 'El texto del botón es obligatorio'
+                } else if (value.length > 20) {
+                    newErrors.buttonText = 'El texto del botón debe tener máximo 20 caracteres'
+                } else {
+                    delete newErrors.buttonText
+                }
+                break
+
             case 'paymentLink':
                 if (!value.trim()) {
-                    newErrors.paymentLink = 'El link de pago es obligatorio'
+                    newErrors.paymentLink = 'El link es obligatorio'
                 } else if (!value.includes('.')) {
                     newErrors.paymentLink = 'Ingresa un link válido'
                 } else {
@@ -103,7 +116,7 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
         return number.toLocaleString('es-CL')
     }
 
-    // Manejar cambios en los campos
+    // Handlers para los campos
     const handleTitleChange = (value: string) => {
         setTitle(value)
         validateField('title', value)
@@ -117,11 +130,21 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
     }
 
     const handlePriceChange = (value: string) => {
-        // Solo permitir números
-        const numbersOnly = value.replace(/\D/g, '')
-        setPrice(numbersOnly)
-        validateField('price', numbersOnly)
-        onUpdate({ price: parseInt(numbersOnly) || 0 })
+        const cleanValue = value.replace(/\D/g, '')
+        setPrice(cleanValue)
+        validateField('price', cleanValue)
+        onUpdate({ price: parseInt(cleanValue) || 0 })
+    }
+
+    const handlePriceVisibleChange = (visible: boolean) => {
+        setPriceVisible(visible)
+        onUpdate({ priceVisible: visible })
+    }
+
+    const handleButtonTextChange = (value: string) => {
+        setButtonText(value)
+        validateField('buttonText', value)
+        onUpdate({ buttonText: value })
     }
 
     const handlePaymentLinkChange = (value: string) => {
@@ -147,68 +170,51 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
         onUpdate({ categories: newCategories })
     }
 
-    // Verificar si el producto está completo
-    const isComplete = title.trim() && 
-                      price && 
-                      paymentLink.trim() && 
-                      description.trim() && 
-                      images.length > 0 &&
+    // Verificar si el item está completo
+    const isComplete = title.trim() && price && buttonText.trim() && paymentLink.trim() && 
+                      description.trim() && images.length > 0 && 
                       Object.keys(errors).length === 0
 
     return (
-        <div className="bg-white rounded-lg shadow-md p-4 mb-4 border-l-4 border-green-500">
-            {/* Header del producto */}
-            <div className="flex justify-between items-center mb-4">
-                <div className="flex items-center space-x-2">
-                    <div className="w-8 h-8 bg-green-500 rounded-full flex items-center justify-center">
-                        <svg className="w-4 h-4 text-white" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z" />
-                        </svg>
-                    </div>
+        <div className="bg-white rounded-lg shadow-md border border-gray-200">
+            {/* Header */}
+            <div className="border-b border-gray-200 p-4">
+                <div className="flex items-center justify-between">
                     <h3 className="text-lg font-semibold text-gray-800">
-                        Producto {isComplete ? '✅' : '⚠️'}
+                        📦 {item ? `Editando Item #${item.id}` : 'Nuevo Item'}
                     </h3>
+                    <button
+                        onClick={onRemove}
+                        className="text-red-600 hover:text-red-800 font-medium text-sm px-3 py-1 rounded hover:bg-red-50 transition-colors"
+                    >
+                        Eliminar
+                    </button>
                 </div>
-                <button
-                    type="button"
-                    onClick={onRemove}
-                    className="w-8 h-8 bg-red-500 hover:bg-red-600 text-white rounded-full flex items-center justify-center transition-colors"
-                    title="Eliminar producto"
-                >
-                    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                    </svg>
-                </button>
             </div>
 
             {/* Carrusel de imágenes */}
-            <div className="mb-6">
-                <label className="block text-sm font-medium text-gray-700 mb-2">
-                    Imágenes del producto *
-                </label>
+            <div className="p-4 border-b border-gray-200">
                 <ImageCarousel
                     images={images}
                     onImagesChange={handleImagesChange}
-                    maxImages={10}
-                    maxSizeInMB={10}
+                    bucketName="items"
+                    folderPrefix={item?.id ? `item-${item.id}` : undefined}
+                    error={errors.images}
                 />
-                {errors.images && (
-                    <p className="mt-1 text-sm text-red-600">{errors.images}</p>
-                )}
             </div>
 
-            {/* Campos del formulario */}
-            <div className="space-y-4">
+            {/* Formulario */}
+            <div className="p-6 space-y-4">
                 {/* Título */}
                 <div>
                     <label className="block text-sm font-medium text-gray-700 mb-1">
-                        Título del producto *
+                        Título *
                     </label>
                     <input
                         type="text"
                         value={title}
                         onChange={(e) => handleTitleChange(e.target.value)}
-                        placeholder="Ej: Jockey Rosado"
+                        placeholder="Ej: Consultora Personalizada"
                         maxLength={50}
                         className={`w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 ${
                             errors.title ? 'border-red-300 focus:ring-red-500' : 'border-gray-300'
@@ -231,7 +237,7 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                         type="text"
                         value={subtitle}
                         onChange={(e) => handleSubtitleChange(e.target.value)}
-                        placeholder="Ej: Edición especial con logo bordado"
+                        placeholder="Ej: Sesión de 1 hora para análisis biomecánico"
                         maxLength={150}
                         className={`w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 ${
                             errors.subtitle ? 'border-red-300 focus:ring-red-500' : 'border-gray-300'
@@ -245,18 +251,39 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                     </div>
                 </div>
 
-                {/* Precio */}
+                {/* Precio con toggle de visibilidad */}
                 <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-1">
-                        Precio (CLP) *
-                    </label>
+                    <div className="flex items-center justify-between mb-1">
+                        <label className="block text-sm font-medium text-gray-700">
+                            Precio (CLP) *
+                        </label>
+                        <div className="flex items-center space-x-2">
+                            <span className="text-xs text-gray-500">Visible:</span>
+                            <button
+                                type="button"
+                                onClick={() => handlePriceVisibleChange(!priceVisible)}
+                                className={`p-1 rounded-md transition-colors ${
+                                    priceVisible 
+                                        ? 'text-green-600 hover:bg-green-50' 
+                                        : 'text-gray-400 hover:bg-gray-50'
+                                }`}
+                                title={priceVisible ? 'Precio visible' : 'Precio oculto'}
+                            >
+                                {priceVisible ? (
+                                    <EyeIcon className="w-5 h-5" />
+                                ) : (
+                                    <EyeSlashIcon className="w-5 h-5" />
+                                )}
+                            </button>
+                        </div>
+                    </div>
                     <div className="relative">
                         <span className="absolute left-3 top-2 text-gray-500">$</span>
                         <input
                             type="text"
                             value={formatPrice(price)}
                             onChange={(e) => handlePriceChange(e.target.value)}
-                            placeholder="15000"
+                            placeholder="25000"
                             className={`w-full pl-8 pr-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 ${
                                 errors.price ? 'border-red-300 focus:ring-red-500' : 'border-gray-300'
                             }`}
@@ -267,21 +294,44 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                     )}
                     {price && !errors.price && (
                         <p className="mt-1 text-sm text-green-600">
-                            Se mostrará como: ${formatPrice(price)}
+                            Se mostrará como: {priceVisible ? `$${formatPrice(price)}` : 'Precio oculto'}
                         </p>
                     )}
                 </div>
 
-                {/* Link de pago */}
+                {/* Texto del botón */}
                 <div>
                     <label className="block text-sm font-medium text-gray-700 mb-1">
-                        Link de pago *
+                        Texto del botón *
+                    </label>
+                    <input
+                        type="text"
+                        value={buttonText}
+                        onChange={(e) => handleButtonTextChange(e.target.value)}
+                        placeholder="Ver más"
+                        maxLength={20}
+                        className={`w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 ${
+                            errors.buttonText ? 'border-red-300 focus:ring-red-500' : 'border-gray-300'
+                        }`}
+                    />
+                    <div className="flex justify-between mt-1">
+                        {errors.buttonText && (
+                            <p className="text-sm text-red-600">{errors.buttonText}</p>
+                        )}
+                        <p className="text-xs text-gray-500 ml-auto">{buttonText.length}/20</p>
+                    </div>
+                </div>
+
+                {/* Link */}
+                <div>
+                    <label className="block text-sm font-medium text-gray-700 mb-1">
+                        Link *
                     </label>
                     <input
                         type="url"
                         value={paymentLink}
                         onChange={(e) => handlePaymentLinkChange(e.target.value)}
-                        placeholder="https://mpago.la/2wPmNL9"
+                        placeholder="https://ejemplo.com/contacto"
                         className={`w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 ${
                             errors.paymentLink ? 'border-red-300 focus:ring-red-500' : 'border-gray-300'
                         }`}
@@ -299,7 +349,7 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                     <textarea
                         value={description}
                         onChange={(e) => handleDescriptionChange(e.target.value)}
-                        placeholder="Describe tu producto en detalle..."
+                        placeholder="Describe tu item en detalle..."
                         maxLength={1000}
                         rows={5}
                         className={`w-full px-3 py-2 border rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 resize-none ${
@@ -331,11 +381,11 @@ export function ProductForm({ product, availableCategories, onUpdate, onRemove }
                 </div>
             </div>
 
-            {/* Estado del producto */}
+            {/* Estado del item */}
             <div className="mt-4 p-3 rounded-lg bg-gray-50">
                 <div className="flex items-center justify-between">
                     <span className="text-sm font-medium text-gray-700">
-                        Estado del producto:
+                        Estado del item:
                     </span>
                     <span className={`text-sm font-semibold ${
                         isComplete ? 'text-green-600' : 'text-orange-600'
