+++ src/components/features/profile/index.ts
@@ -1,0 +1,4 @@
+// Profile Configuration Components
+export { BackgroundConfig } from './BackgroundConfig'
+export { SocialIconsConfig } from './SocialIconsConfig'
+export { StyleConfig } from './StyleConfig'
