+++ src/components/index.ts
@@ -1,30 +1,8 @@
-// Componentes principales
-export { AvatarUpload } from './AvatarUpload'
-export { FileInfo } from './FileInfo'
-export { SocialIconsConfig } from './SocialIconsConfig'
-export { BackgroundConfig } from './BackgroundConfig'
-export { StyleConfig } from './StyleConfig'
-export { UnsavedChangesIndicator } from './UnsavedChangesIndicator'
-
-// Componentes del formulario
-export { LinkCard } from './SortableLinksForm/LinkCard'
-export { DeleteModal } from './SortableLinksForm/DeleteModal'
-export { ActionButtons } from './SortableLinksForm/ActionButtons'
-
-// Componentes de productos
-export { ProductForm } from './ProductForm'
-export { ProductModal } from './ProductModal'
-export { ImageCarousel } from './ImageCarousel'
-
-// Componentes de items
-export { ItemForm } from './ItemForm'
-export { ItemCard } from './ItemCard'
-export { ItemModal } from './ItemModal'
-
-// Componentes de categorías
-export { default as CategorySelector } from './CategorySelector'
-export { default as CategoryManager } from './CategoryManager'
-export { default as CategoryFilter } from './CategoryFilter'
-
-// Componente principal con productos
-export { SortableLinksFormWithProducts } from './SortableLinksFormWithProducts'
+// Re-export from organized structure
+export * from './common/ui'
+export * from './common/forms'
+export * from './common/media'
+export * from './features/categories'
+export * from './features/products'
+export * from './features/profile'
+export * from './features/links'
